+++ src/handlers/error.js
@@ -4,10 +4,9 @@ module.exports = {
 
 function handle(err, req, res, next) {
   req.error = err;
-  console.log('ERROR HANDLE');
-
   res.status(500).send({
     status: 500,
-    message: err.message
+    message: 'An error is happened',
+    stack: err.message
   });
 }